src/third_party/mozjs/extract/intl/uconv/tests/test_bug335816.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=335816
-->
<head>
  <meta http-equiv="Content-type" content="text/html; charset=UTF-8">
  <title>Test for Bug 335816</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="test()">
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 335816 **/
function test()
{
  ok($("display").innerHTML != "Keep that breathless charm", "Hidden script not executed");
  SimpleTest.finish();
}

function WontYouPleaseArrangeIt()
{
  $("display").innerHTML = "Keep that breathless charm";
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=335816">Mozilla Bug 335816</a>
<p id="display">Lovely ... Never, ever change.</p>

<!-- There is a UTF-8 BOM just here - - - 
                                         |  which should not be stripped 
                                         v      -->
<div id="content" style="display: none"><﻿script>WontYouPleaseArrangeIt();</script></div>
</body>
</html>
